chore: breakup bundle tests
