+++ .github/workflows/parallel-tests.yaml
@@ -32,6 +32,9 @@ jobs:
   test-bundle:
     runs-on: ubuntu-latest
     needs: build
+    strategy:
+      matrix:
+        test: [flags, inspect, deploy, create, misc]
     steps:
       - name: Checkout
         uses: actions/checkout@692973e3d937129bcbf40652eb9f2f61becf3332 # v4.1.7
@@ -43,7 +46,7 @@ jobs:
 
       - name: Run e2e bundle tests
         run: |
-          build/uds run test:bundle
+          build/uds run test:bundle-${{ matrix.test }}
 
       - name: Save logs
         if: always()
