+++ src/test/packages/podinfo-nginx/zarf.yaml
@@ -9,7 +9,7 @@ components:
     import:
       path: ../podinfo
 
-  - name: nginx-remote
+  - name: nginx
     import:
       path: ../nginx
